// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const BookStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  CANCELLED: 'CANCELLED'
} as const

export type BookStatus = (typeof BookStatus)[keyof typeof BookStatus]


export const Rating = {
  I: 'I',
  II: 'II',
  III: 'III',
  IIII: 'IIII',
  IIIII: 'IIIII',
  IIIIII: 'IIIIII',
  IIIIIII: 'IIIIIII',
  IIIIIIII: 'IIIIIIII',
  IIIIIIIII: 'IIIIIIIII',
  IIIIIIIIII: 'IIIIIIIIII'
} as const

export type Rating = (typeof Rating)[keyof typeof Rating]


export const Otype = {
  MOVIE: 'MOVIE',
  FOOD: 'FOOD',
  SNACKS: 'SNACKS',
  COMBO: 'COMBO'
} as const

export type Otype = (typeof Otype)[keyof typeof Otype]